+++ src/explorer/items/libraryItem.ts
@@ -31,9 +31,9 @@ export function createLibraryItem({ uri, configSource }: MacroLibrary) {
 
   function updateUntitledLibraryItem(item: vscode.TreeItem) {
     item.contextValue = 'macroLibrary,untitled';
-    item.iconPath = new vscode.ThemeIcon('server-process');
+    item.iconPath = new vscode.ThemeIcon('root-folder');
     item.tooltip = new vscode.MarkdownString(
-      'This library includes any untitled macro editors until you save them.  \nUse `.macro.js` or `.macro.ts` extension to enable IntelliSense  \nand macro tooling.',
+      'This library includes all `untitled` macro editors.  \n Save them to a library, or use `.macro.js` or `.macro.ts`  \nas file extension to enable IntelliSense and macro tooling.',
     );
   }
 }